requirements.txt
flask-sqlalchemy
pymysql  # Alternative MySQL adapter
mysql-connector-python
